+++ pkg/accounting/tally/tally.go
@@ -12,10 +12,10 @@ import (
 	"go.uber.org/zap"
 
 	"storj.io/storj/pkg/accounting"
+	"storj.io/storj/pkg/accounting/live"
 	"storj.io/storj/pkg/overlay"
 	"storj.io/storj/pkg/pb"
 	"storj.io/storj/pkg/storj"
-	"storj.io/storj/satellite/liveaccounting"
 	"storj.io/storj/satellite/metainfo"
 	"storj.io/storj/storage"
 )
@@ -33,11 +33,11 @@ type Service struct {
 	limit          int
 	ticker         *time.Ticker
 	accountingDB   accounting.DB
-	liveAccounting liveaccounting.Service
+	liveAccounting live.Service
 }
 
 // New creates a new tally Service
-func New(logger *zap.Logger, accountingDB accounting.DB, liveAccounting liveaccounting.Service, metainfo *metainfo.Service, overlay *overlay.Cache, limit int, interval time.Duration) *Service {
+func New(logger *zap.Logger, accountingDB accounting.DB, liveAccounting live.Service, metainfo *metainfo.Service, overlay *overlay.Cache, limit int, interval time.Duration) *Service {
 	return &Service{
 		logger:         logger,
 		metainfo:       metainfo,
@@ -68,7 +68,7 @@ func (t *Service) Run(ctx context.Context) (err error) {
 
 // Tally calculates data-at-rest usage once
 func (t *Service) Tally(ctx context.Context) error {
-	// The liveaccounting store will only keep a delta to space used relative
+	// The live accounting store will only keep a delta to space used relative
 	// to the latest tally. Since a new tally is beginning, we will zero it out
 	// now. There is a window between this call and the point where the tally DB
 	// transaction starts, during which some changes in space usage may be
